+++ internal/checks/base.go
@@ -31,6 +31,7 @@ var (
 		VectorMatchingCheckName,
 		CostCheckName,
 		CounterCheckName,
+		PerformanceCheckName,
 		SeriesCheckName,
 		RuleDependencyCheckName,
 		RuleDuplicateCheckName,